Added authorization policy.
This means that we need to actually have the application in the JWT response before we show the secure page.

If not, they are bounced to a non existent page, 'Account/AccessDenied'.

This shouldn't be an issue for folks going through the previous tutorial because we make sure to register the user for the application.